# models/book.go
package models

// Book represents book info
// @Description Book data structure
type Book struct {
	ID       string `json:"id"`
	Title    string `json:"title"`
	Author   string `json:"author"`
	Quantity int    `json:"quantity"`
}